refactor: move iterators to their own file
